+++ packages/druxt/src/nuxt/index.js
@@ -1,6 +1,6 @@
 import chalk from 'chalk'
 import { join, resolve } from 'path'
-import meta from '../package.json'
+import meta from '../../package.json'
 
 /**
  * Nuxt module function to install Druxt.
@@ -86,6 +86,23 @@ const DruxtNuxtModule = function (moduleOptions = {}) {
     options
   })
 
+  // Make the $druxt plugin the first to load.
+  const extendPlugins = this.options.extendPlugins
+  this.options.extendPlugins = (plugins) => {
+    // Run the user defined extendPlugins function if defined.
+    plugins = typeof extendPlugins === 'function' ? extendPlugins(plugins) : plugins
+
+    // Find the $druxt plugin.
+    const index = plugins.findIndex(({ src }) => src === `${this.options.buildDir}/druxt.js`)
+    const plugin = plugins[index]
+
+    // Move the $druxt plugin to first place.
+    plugins.splice(index, 1)
+    plugins.unshift(plugin)
+
+    return plugins
+  }
+
   // Add Vuex plugin.
   this.addPlugin({
     src: resolve(__dirname, '../templates/store.js'),
@@ -103,8 +120,6 @@ const DruxtNuxtModule = function (moduleOptions = {}) {
   this.options.cli.badgeMessages.push(`${chalk.blue.bold('Druxt')} @ v${meta.version}`)
   this.options.cli.badgeMessages.push(`${chalk.bold('API:')} ${chalk.blue.underline(options.baseUrl + options.endpoint)}`)
 }
-
-DruxtNuxtModule.meta = meta
 
 export { DruxtNuxtModule }