+++ packages/druxt/test/components/DruxtModule.test.js
@@ -64,7 +64,7 @@ describe('DruxtModule component', () => {
       components: { DruxtModule },
       data: () => ({ model: null })
     }
-    const wrapper = mount(Component, { localVue, mocks, stubs: ['DruxtWrapper'] })
+    const wrapper = mount(Component, { localVue, mocks, stubs: ['DruxtDebug'] })
 
     // Default state.
     expect(wrapper.vm.model).toStrictEqual(null)
@@ -79,6 +79,59 @@ describe('DruxtModule component', () => {
     expect(wrapper.vm.$refs.module.component.props.value).toStrictEqual({ test: true })
     expect(wrapper.vm.$refs.module.model).toStrictEqual({ test: true })
     expect(wrapper.vm.$refs.module.value).toStrictEqual({ test: true })
+  })
+
+  test('error', async () => {
+    const CustomModule = {
+      name: 'CustomModule',
+      extends: DruxtModule,
+    }
+
+    let wrapper = mount(CustomModule, {localVue, mocks, stubs: ['DruxtDebug'] })
+    await wrapper.vm.$options.fetch.call(wrapper.vm)
+
+    expect(wrapper.vm.component.is).toBe('DruxtWrapper')
+
+    // Simulate error.
+    const err = {
+      response: {
+        status: '500',
+        statusText: 'Test error',
+        data: {
+          errors: [{
+            details: 'Test error'
+          }]
+        }
+      }
+    }
+    wrapper.vm.error(err)
+
+    // Expect the component to be DruxtDebug.
+    expect(wrapper.vm.component.is).toBe('DruxtDebug')
+    // Expect the component props to match snapshot.
+    expect(wrapper.vm.component.props).toMatchSnapshot()
+
+    // Fetch config error.
+    CustomModule.druxt = {
+      fetchConfig: async () => { throw new Error('Fetch config error') }
+    }
+    wrapper = mount(CustomModule, {localVue, mocks, stubs: ['DruxtDebug'] })
+    await wrapper.vm.$options.fetch.call(wrapper.vm)
+    // Expect the component to be DruxtDebug.
+    expect(wrapper.vm.component.is).toBe('DruxtDebug')
+    // Expect the component props to match snapshot.
+    expect(wrapper.vm.component.props).toMatchSnapshot()
+
+    // Fetch data error.
+    CustomModule.druxt = {
+      fetchData: async () => { throw new Error('Fetch data error') }
+    }
+    wrapper = mount(CustomModule, {localVue, mocks, stubs: ['DruxtDebug'] })
+    await wrapper.vm.$options.fetch.call(wrapper.vm)
+    // Expect the component to be DruxtDebug.
+    expect(wrapper.vm.component.is).toBe('DruxtDebug')
+    // Expect the component props to match snapshot.
+    expect(wrapper.vm.component.props).toMatchSnapshot()
   })
 
   test('custom module - no wrapper', async () => {